requirements.txt
tensorflow==2.13.0
streamlit==1.28.0
pillow==10.0.1
numpy==1.24.3
matplotlib==3.7.2